refactor: Refactoring habit section and habit page
