+++ src/hooks/useHabitMutations/types.ts
@@ -1,0 +1,6 @@
+import { HabitUpdateInput } from "@/api/generated";
+
+export interface HabitUpdateInputParameter {
+  id: number;
+  update: HabitUpdateInput;
+}
