+++ src/pages/MainPage/Habit/HabitSection.tsx
@@ -1,80 +1,67 @@
-import { useMutation } from "@tanstack/react-query";
-
-import { HabitPublic, HabitWithLog } from "@/api/generated";
-import EmptyCard from "@/components/EmptyCard";
+import { HabitWithLog } from "@/api/generated";
 import {
   convertMinutesToHours,
   getFormatMinutesWithMeridiem,
   getTimeDifferenceFromNow,
+  parseRepeatDays,
 } from "@/utils/time";
-import { habitsApi } from "@/api/client";
-import { useMessageStore } from "@/state/useMessageStore";
-import { HabitModalSubmitProps } from "@/components/HabitModal/types";
 import HabitModal from "@/components/HabitModal";
-import useModalWithState from "@/hooks/useModalWithState";
-import useModal from "@/hooks/useModal";
-
-import { HabitUpdateInputParameter } from "../types";
+import HabitInformation from "@/components/HabitInformation";
+import useHabitMutations from "@/hooks/useHabitMutations";
 
 interface Props {
   habitList: Array<HabitWithLog>;
-  fetchData: () => void;
+  reloadHabitList: () => void;
 
   fullData?: boolean;
 }
 
-export default function HabitSection({ habitList, fetchData }: Props) {
-  const addMessage = useMessageStore((state) => state.addMessage);
-  const achieveHabitMutation = useMutation({
-    mutationFn: habitsApi.achieveHabit,
-    onSuccess: () => {
-      fetchData();
-    },
-    onError: () => {
-      addMessage({
-        message: "습관 달성 요청에 실패했습니다.",
-        type: "error",
-      });
-    },
-  });
-
+export default function HabitSection({ habitList, reloadHabitList }: Props) {
+  const {
+    createHabitModal,
+    updateHabitModal,
+    achieveHabitMutation,
+    deleteHabitMutation,
+    createHabit,
+    updateHabit,
+    deleteHabit,
+  } = useHabitMutations({ reloadHabitList });
   const {
     modalState: selectedHabit,
     openModal: openUpdateModal,
     closeModal: closeUpdateModal,
     modalRef: updateModalRef,
-  } = useModalWithState<HabitPublic>();
+  } = updateHabitModal;
   const {
-    isModalOpened: isCreateModalOpened,
     closeModal: closeCreateModal,
     openModal: openCreateModal,
     modalRef: createModalRef,
-  } = useModal();
+    isModalOpened: isCreateModalOpened,
+  } = createHabitModal;
+
+  const renderUncompletedHabitCount = (habit: HabitWithLog, count: number) => {
+    return (
+      <>
+        <p>
+          {habit.log_list.length
+            ? `${convertMinutesToHours(getTimeDifferenceFromNow(habit.log_list[0].completed_at))} 전에 실천했어요.`
+            : `${getFormatMinutesWithMeridiem(habit.start_time_minutes)} 습관 시작이에요.`}
+        </p>
+
+        <div>
+          <p className="text-primary text-xs mt-1">
+            {habit.log_list.length > 0
+              ? `오늘 총 ${count}번 중 ${habit.log_list.length}번 완료!`
+              : "지금 습관을 실천해보세요!"}
+          </p>
+        </div>
+      </>
+    );
+  };
 
   const renderHabitList = (list: Array<HabitWithLog>) => {
-    if (list.length === 0) {
-      return (
-        <EmptyCard label="습관">
-          <div>
-            <div className="text-sm text-gray-500">
-              <p>
-                습관은 <b>하루에 여러번 반복되는 일</b>을<br /> 편리하게 관리할
-                수 있어요.
-              </p>
-              <p className="mt-1">
-                주기적으로 반복되는 할 일이라고
-                <br /> 생각하면 편해요.
-              </p>
-            </div>
-            <button
-              className="btn btn-primary btn-outline mt-2"
-              onClick={openCreateModal}
-            >
-              습관 추가하기
-            </button>
-          </div>
-        </EmptyCard>
-      );
+    if (!list.length) {
+      return <HabitInformation onModalOpen={openCreateModal} />;
     }
 
     return list.map((data) => {
@@ -109,26 +96,15 @@ export default function HabitSection({ habitList, fetchData }: Props) {
                   {data.title}
                 </h2>
               </div>
+
               {done ? (
                 <p className="text-primary text-xs mt-1">
                   오늘 총 {count}번의 습관을 모두 완료했어요!
                 </p>
               ) : (
-                <>
-                  <p>
-                    {data.log_list.length
-                      ? `${convertMinutesToHours(getTimeDifferenceFromNow(data.log_list[0].completed_at))} 전에 실천했어요.`
-                      : `${getFormatMinutesWithMeridiem(data.start_time_minutes)} 습관 시작이에요.`}
-                  </p>
-                  <div>
-                    <p className="text-primary text-xs mt-1">
-                      {data.log_list.length > 0
-                        ? `오늘 총 ${count}번 중 ${data.log_list.length}번 완료!`
-                        : "지금 습관을 실천해보세요!"}
-                    </p>
-                  </div>
-                </>
+                renderUncompletedHabitCount(data, count)
               )}
+
               {data.log_list.length ? (
                 <progress
                   className="progress progress-primary w-56"
@@ -139,6 +115,7 @@ export default function HabitSection({ habitList, fetchData }: Props) {
                 <></>
               )}
             </div>
+
             <input
               type="checkbox"
               disabled={done}
@@ -152,108 +129,6 @@ export default function HabitSection({ habitList, fetchData }: Props) {
         </li>
       );
     });
-  };
-
-  const createHabitMutation = useMutation({
-    mutationFn: habitsApi.createHabit,
-    onSuccess: () => {
-      closeCreateModal();
-      fetchData();
-    },
-    onError: () => {
-      addMessage({
-        message: "습관 추가에 실패했습니다.",
-        type: "error",
-      });
-    },
-  });
-
-  const handleAddHabitSubmit = (params: HabitModalSubmitProps) => {
-    createHabitMutation.mutate({
-      title: params.title,
-      start_time_minutes: params.startTimeMinutes,
-      end_time_minutes: params.endTimeMinutes,
-      repeat_days: parseRepeatDaysToServerFormat(params.repeatDays),
-      repeat_time_minutes: params.repeatIntervalMinutes,
-    });
-  };
-
-  const updateHabitMutation = useMutation({
-    mutationFn: (input: HabitUpdateInputParameter) =>
-      habitsApi.updateHabit(input.id, input.update),
-    onSuccess: () => {
-      closeUpdateModal();
-      fetchData();
-    },
-    onError: () => {
-      addMessage({
-        message: "습관 수정에 실패했습니다.",
-        type: "error",
-      });
-    },
-  });
-
-  const handleUpdateHabitSubmit = (params: HabitModalSubmitProps) => {
-    if (!selectedHabit) {
-      return;
-    }
-
-    updateHabitMutation.mutate({
-      id: selectedHabit.id,
-      update: {
-        title: params.title,
-        start_time_minutes: params.startTimeMinutes,
-        end_time_minutes: params.endTimeMinutes,
-        repeat_days: parseRepeatDaysToServerFormat(params.repeatDays),
-        repeat_time_minutes: params.repeatIntervalMinutes,
-        activated: true,
-      },
-    });
-  };
-
-  const parseRepeatDays = (repeatDays: Array<number>): Array<number> => {
-    const result = Array.from({ length: 7 }, () => 0);
-
-    repeatDays.forEach((value) => {
-      result[value] = 1;
-    });
-
-    return result;
-  };
-
-  const parseRepeatDaysToServerFormat = (
-    repeatDays: Array<number>
-  ): Array<number> => {
-    const result: Array<number> = [];
-
-    repeatDays.forEach((value, i) => {
-      if (value) {
-        result.push(i);
-      }
-    });
-
-    return result;
-  };
-
-  const deleteHabitMutation = useMutation({
-    mutationFn: habitsApi.deleteHabit,
-    onSuccess: () => {
-      closeUpdateModal();
-      fetchData();
-      addMessage({
-        message: "습관을 삭제했습니다.",
-      });
-    },
-    onError: () => {
-      addMessage({
-        message: "습관 삭제에 실패했습니다.",
-        type: "error",
-      });
-    },
-  });
-
-  const handleHabitDeleteClick = (id: number) => {
-    deleteHabitMutation.mutate(id);
   };
 
   return (
@@ -274,16 +149,16 @@ export default function HabitSection({ habitList, fetchData }: Props) {
         key={isCreateModalOpened ? "open" : "close"}
         modalRef={createModalRef}
         onCancel={closeCreateModal}
-        onSubmit={handleAddHabitSubmit}
+        onSubmit={createHabit}
         modalId="habit-add-modal"
         modalTitle="습관 추가하기"
       />
 
-      {selectedHabit ? (
+      {selectedHabit && (
         <HabitModal
           modalRef={updateModalRef}
           onCancel={closeUpdateModal}
-          onSubmit={handleUpdateHabitSubmit}
+          onSubmit={updateHabit}
           modalId="habit-add-modal"
           modalTitle="습관 수정하기"
           title={selectedHabit.title}
@@ -294,15 +169,13 @@ export default function HabitSection({ habitList, fetchData }: Props) {
           extraButton={
             <button
               disabled={deleteHabitMutation.isPending}
-              onClick={() => handleHabitDeleteClick(selectedHabit.id)}
+              onClick={() => deleteHabit(selectedHabit.id)}
               className="btn btn-error"
             >
               삭제하기
             </button>
           }
         />
-      ) : (
-        <></>
       )}
     </>
   );
